mtp-server: don't throw error if /dev/mtp_usb is unavailable
On some devices, e.g. Volla 22, sometimes /dev/mtp_usb is not available.
If this is the case and mtp-server.service tries to start, it will error
out immediately and keep restarting the service indefinitely, causing
unnecessary spam in the journal.
Now instead of throwing an error, just wait for /dev/mtp_usb to appear
and then continue as normal when it does.